Add zero value and reorganize the project
