+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"D:\\webprojects\\react-redux-practice\\src\\index.js":"1","D:\\webprojects\\react-redux-practice\\src\\App.js":"2","D:\\webprojects\\react-redux-practice\\src\\serviceWorker.js":"3","D:\\webprojects\\react-redux-practice\\src\\app\\store.js":"4","D:\\webprojects\\react-redux-practice\\src\\features\\counter\\counterSlice.js":"5","D:\\webprojects\\react-redux-practice\\src\\features\\counter\\Counter.js":"6"},{"size":644,"mtime":1608975822000,"results":"7","hashOfConfig":"8"},{"size":1463,"mtime":1608975822000,"results":"9","hashOfConfig":"8"},{"size":5141,"mtime":1608975822000,"results":"10","hashOfConfig":"8"},{"size":286,"mtime":1609193698098,"results":"11","hashOfConfig":"8"},{"size":809,"mtime":1609193799867,"results":"12","hashOfConfig":"8"},{"size":1772,"mtime":1609193790916,"results":"13","hashOfConfig":"8"},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},"fev7de",{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"19","messages":"20","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"D:\\webprojects\\react-redux-practice\\src\\index.js",[],["27","28"],"D:\\webprojects\\react-redux-practice\\src\\App.js",[],"D:\\webprojects\\react-redux-practice\\src\\serviceWorker.js",[],"D:\\webprojects\\react-redux-practice\\src\\app\\store.js",[],"D:\\webprojects\\react-redux-practice\\src\\features\\counter\\counterSlice.js",[],"D:\\webprojects\\react-redux-practice\\src\\features\\counter\\Counter.js",[],{"ruleId":"29","replacedBy":"30"},{"ruleId":"31","replacedBy":"32"},"no-native-reassign",["33"],"no-negated-in-lhs",["34"],"no-global-assign","no-unsafe-negation"]
+[{"D:\\webprojects\\react-redux-practice\\src\\index.js":"1","D:\\webprojects\\react-redux-practice\\src\\App.js":"2","D:\\webprojects\\react-redux-practice\\src\\serviceWorker.js":"3","D:\\webprojects\\react-redux-practice\\src\\app\\store.js":"4","D:\\webprojects\\react-redux-practice\\src\\features\\counter\\Counter.js":"5","D:\\webprojects\\react-redux-practice\\src\\features\\counter\\counterReducer.js":"6","D:\\webprojects\\react-redux-practice\\src\\features\\counter\\actions.js":"7"},{"size":644,"mtime":1608975822000,"results":"8","hashOfConfig":"9"},{"size":1463,"mtime":1608975822000,"results":"10","hashOfConfig":"9"},{"size":5141,"mtime":1608975822000,"results":"11","hashOfConfig":"9"},{"size":289,"mtime":1609232336719,"results":"12","hashOfConfig":"9"},{"size":1994,"mtime":1609233259694,"results":"13","hashOfConfig":"9"},{"size":604,"mtime":1609233588227,"results":"14","hashOfConfig":"9"},{"size":322,"mtime":1609233183807,"results":"15","hashOfConfig":"9"},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},"fev7de",{"filePath":"19","messages":"20","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"D:\\webprojects\\react-redux-practice\\src\\index.js",[],["31","32"],"D:\\webprojects\\react-redux-practice\\src\\App.js",[],"D:\\webprojects\\react-redux-practice\\src\\serviceWorker.js",[],"D:\\webprojects\\react-redux-practice\\src\\app\\store.js",[],"D:\\webprojects\\react-redux-practice\\src\\features\\counter\\Counter.js",[],"D:\\webprojects\\react-redux-practice\\src\\features\\counter\\counterReducer.js",[],"D:\\webprojects\\react-redux-practice\\src\\features\\counter\\actions.js",[],{"ruleId":"33","replacedBy":"34"},{"ruleId":"35","replacedBy":"36"},"no-native-reassign",["37"],"no-negated-in-lhs",["38"],"no-global-assign","no-unsafe-negation"]
