+++ src/app/store.js
@@ -1,7 +1,7 @@
 import { createStore } from 'redux';
-import counterReducer from '../features/counter/counterSlice';
+import counterReducer from '../features/counter/counterReducer';
 
-let store = createStore(counterReducer)
+let store = createStore(counterReducer);
 
 store.subscribe(() => {
     localStorage.clear();
