+++ src/features/counter/Counter.module.css
@@ -17,6 +17,7 @@
 }
 
 .button {
+  margin: 0 5px;
   appearance: none;
   background: none;
   font-size: 32px;
